# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Contract Management System API
  description: |
    Complete API documentation for the Contract Management System.
    
    ## Authentication
    All API requests require authentication using JWT tokens.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Anonymous requests: 20/hour
    - Authenticated requests: 1000/hour
    - Premium users: 5000/hour
    
    ## WebSocket Support
    Real-time updates are available via WebSocket connections at:
    - Production: wss://api.contractmanagement.com/ws
    - Development: ws://localhost:8000/ws
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@contractmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.contractmanagement.com/v1
    description: Production server
  - url: https://staging-api.contractmanagement.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
        401:
          $ref: '#/components/responses/UnauthorizedError'
...